+++ src/main/java/com/mcmiddleearth/entities/entities/McmeEntity.java
@@ -56,16 +56,16 @@ public interface McmeEntity extends Entity {
 
     public EntityBoundingBox getBoundingBox();
 
-    public int getHealth();
-    public void damage(int damage);
-    public void heal(int damage);
+    public double getHealth();
+    public void damage(double damage);
+    public void heal(double damage);
     public boolean isDead();
 
     public boolean isTerminated();
 
     public void playAnimation(ActionType type);
 
-    public void receiveAttack(McmeEntity damager, int damage, float knockDownFactor);
+    public void receiveAttack(McmeEntity damager, double damage, double knockDownFactor);
     public void attack(McmeEntity target);
 
     public Set<McmeEntity> getEnemies();
